Title: 1, 2, 5 더하기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240409/1, 2, 5 더하기/1-2-5-plus.cpp
@@ -1,17 +1,17 @@
 #include<iostream>
 using namespace std;
-int dp[1000];
+long long dp[1000];
 
 int main() {
 	int s[3] = { 1,2,5 };
-	int n;
+	long long n;
 	dp[1] = 1;
 	dp[2] = 1;
 	dp[5] = 1;
 	
 	cin >> n;
 	
-	for (int i = 1; i <= n; i++) {
+	for (long long i = 1; i <= n; i++) {
 
 		if (dp[i] != 0) {
